backend/blog/views.py
from rest_framework import viewsets
from rest_framework import permissions
from .serializers import PostSerializer
from .models import Post
from rest_framework import generics

class PostList(generics.ListCreateAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer


class PostDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer
