backend/requirements.txt
asgiref==3.6.0
charset-normalizer==2.1.1
Django==4.1.5
django-environ==0.9.0
django-restframework==0.0.1
idna==3.4
oauthlib==3.2.2
pycparser==2.21
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13